+++ pytorch_lightning/accelerators/legacy/ddp2_accelerator.py
@@ -197,8 +197,7 @@ class DDP2Accelerator(Accelerator):
 
         self.trainer.setup_trainer(model)
 
-        # train or test
-        results = self.train_or_test()
+        results = self.train_or_test_or_predict()
 
         # clean up memory
         torch.cuda.empty_cache()