+++ pytorch_lightning/accelerators/legacy/ddp_accelerator.py
@@ -306,8 +306,7 @@ class DDPAccelerator(Accelerator):
         self.barrier('ddp_setup')
         self.trainer.setup_trainer(model)
 
-        # train or test
-        results = self.train_or_test()
+        results = self.train_or_test_or_predict()
 
         # clean up memory
         torch.cuda.empty_cache()