+++ pytorch_lightning/accelerators/legacy/ddp_cpu_spawn_accelerator.py
@@ -152,9 +152,7 @@ class DDPCPUSpawnAccelerator(Accelerator):
         model = self.configure_ddp(model, device_ids)
 
         self.trainer.setup_trainer(model)
-
-        # train or test
-        results = self.train_or_test()
+        results = self.train_or_test_or_predict()
 
         # get original model
         model = self.trainer.get_model()