+++ pytorch_lightning/accelerators/legacy/horovod_accelerator.py
@@ -105,8 +105,7 @@ class HorovodAccelerator(Accelerator):
 
             self.trainer.setup_trainer(self.trainer.model)
 
-            # train or test
-            results = self.train_or_test()
+            results = self.train_or_test_or_predict()
 
         # Make sure all workers have finished training before returning to the user
         hvd.join()