+++ pytorch_lightning/accelerators/legacy/tpu_accelerator.py
@@ -134,8 +134,7 @@ class TPUAccelerator(Accelerator):
 
         self.trainer.setup_trainer(model)
 
-        # train or test
-        results = self.train_or_test()
+        results = self.train_or_test_or_predict()
 
         # save weights at the end of training
         self.__save_end_of_training_weights(model, trainer)