+++ pytorch_lightning/plugins/training_type/single_tpu.py
@@ -43,7 +43,7 @@ class SingleTPUPlugin(SingleDevicePlugin):
         self.tpu_local_core_rank = xm.get_local_ordinal()
         self.tpu_global_core_rank = xm.get_ordinal()
 
-    def post_training(self) -> None:
+    def post_dispatch(self) -> None:
         model = self.lightning_module
 
         if on_colab_kaggle():