+++ pytorch_lightning/plugins/training_type/training_type_plugin.py
@@ -112,11 +112,15 @@ class TrainingTypePlugin(Plugin, ABC):
 
     def start_training(self, trainer: 'Trainer') -> None:
         # double dispatch to initiate the training loop
-        self._results = trainer.train()
+        self._results = trainer.run_train()
 
     def start_testing(self, trainer: 'Trainer') -> None:
         # double dispatch to initiate the test loop
         self._results = trainer.run_test()
+
+    def start_predicting(self, trainer: 'Trainer') -> None:
+        # double dispatch to initiate the predicting loop
+        self._results = trainer.run_predict()
 
     def training_step(self, *args, **kwargs):
         return self.lightning_module.training_step(*args, **kwargs)